Productos y categorias

+++ JardinesEF.Datos.Comun/Facades/IProductosRepositorio.cs
@@ -13,7 +13,7 @@ namespace JardinesEF.Datos.Comun.Facades
         int GetCantidad(Func<Producto, bool> predicate);
         List<Producto> Find(Func<Producto, bool> predicate, int cantidadPorPagina, int paginaActual);
         List<Producto> GetLista(int categoriaId);
-        void SetearReservarProducto(int productoId, int cantidad);
-        void ActualizarStock(int productoId, int cantidad);
+        void SetearReservarProducto(int productoId, decimal cantidad);
+        void ActualizarStock(int productoId, decimal cantidad);
     }
 }
